# openapi/platform-v1.yaml
openapi: 3.1.0
info:
  title: ArchGuard Platform API
  version: 1.2.0-platform
  description: Platform MVP control plane secured by an OIDC JWT resource server.
servers:
  - url: /
security:
  - platformIdentity: []
paths:
  /api/v1/projects:
    get:
      operationId: listProjects
      summary: List projects visible to the caller
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: A page of projects for which the caller is a member
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createProject
      summary: Create a project and make the caller its maintainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: getProject
      summary: Get a project visible to the caller
      responses:
        '200':
          description: Project found and caller is a member
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
    patch:
      operationId: updateProject
      summary: Update a project name as a maintainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated and its version incremented
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteProject
      summary: Delete a project and its memberships as a maintainer
      parameters:
        - name: version
          in: query
          required: true
          description: Expected project version for optimistic concurrency
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '204':
          description: Project and memberships deleted; audit records are retained
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/projects/{projectId}/repositories/{repositoryId}/rule-sets:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RepositoryId'
    get:
      operationId: listRuleSets
      responses:
        '200':
          description: Rule sets
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/RuleSet'}
    post:
      operationId: createRuleSet
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateRuleSetRequest'}
      responses:
        '201':
          description: Rule set created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RuleSet'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/projects/{projectId}/repositories/{repositoryId}/rule-sets/{ruleSetId}/versions:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RepositoryId'
      - name: ruleSetId
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: listRuleSetVersions
      responses:
        '200':
          description: Immutable rule set versions
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/RuleSetVersion'}
    post:
      operationId: createRuleSetVersion
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateRuleSetVersionRequest'}
      responses:
        '201':
          description: Validated rule set version
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RuleSetVersion'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/projects/{projectId}/scan-jobs:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: listScanJobs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Scan job history
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ScanJob'}
    post:
      operationId: submitScanJob
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 128}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SubmitScanJobRequest'}
      responses:
        '201':
          description: Scan job queued or existing identical job returned
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScanJob'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/projects/{projectId}/scan-jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/JobId'
    get:
      operationId: getScanJob
      responses:
        '200':
          description: Scan job state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScanJob'}
        '404': {$ref: '#/components/responses/ProjectNotFound'}
  /api/v1/projects/{projectId}/scan-jobs/{jobId}/cancel:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/JobId'
    post:
      operationId: cancelScanJob
      responses:
        '200':
          description: Cancellation requested or queued job cancelled
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScanJob'}
  /api/v1/projects/{projectId}/scan-jobs/{jobId}/result:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/JobId'
    get:
      operationId: getScanResult
      responses:
        '200':
          description: Original validated canonical Scanner report
          content:
            application/json:
              schema: {type: object}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/projects/{projectId}/scan-jobs/{jobId}/findings:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/JobId'
    get:
      operationId: listFindings
      responses:
        '200':
          description: Normalized findings
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Finding'}
  /api/v1/projects/{projectId}/scan-jobs/{jobId}/evidences/{evidenceId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/JobId'
      - name: evidenceId
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getEvidence
      responses:
        '200':
          description: Evidence without source content
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Evidence'}
  /api/v1/projects/{projectId}/scan-jobs/{jobId}/findings/{findingId}/disposition:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/FindingId'
    put:
      operationId: setFindingDisposition
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SetDispositionRequest'}
      responses:
        '200':
          description: Finding disposition updated and audited
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Finding'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/projects/{projectId}/scan-jobs/{jobId}/findings/{findingId}/dispositions:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/FindingId'
    get:
      operationId: listFindingDispositions
      responses:
        '200':
          description: Append-only disposition history
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Disposition'}
...
components:
  securitySchemes:
    platformIdentity:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC JWT with a UUID sub, the configured audience, and permissions in scope/scp.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema: {type: string, format: uuid}
    JobId:
      name: jobId
      in: path
      required: true
      schema: {type: string, format: uuid}
    FindingId:
      name: findingId
      in: path
      required: true
      schema: {type: string, format: uuid}
  headers:
    RequestId:
      description: Lowercase 32-character hexadecimal request correlation identifier
      schema:
        type: string
        pattern: '^[0-9a-f]{32}$'
  responses:
    BadRequest:
      description: Invalid or malformed request
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: OIDC bearer token is missing or invalid
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Caller is a member but lacks the required role or permission
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ProjectNotFound:
      description: Project does not exist or is not visible to the caller
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Unique key, optimistic version, or last-maintainer invariant conflict
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [key, name]
      properties:
        key:
          type: string
          pattern: '^[a-z][a-z0-9-]{2,62}$'
        name:
          type: string
          minLength: 1
          maxLength: 120
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      required: [name, version]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        version:
          type: integer
          format: int64
          minimum: 0
...
    Project:
      type: object
      additionalProperties: false
      required: [id, key, name, createdBy, createdAt, version]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 0
    ProjectPage:
      type: object
      additionalProperties: false
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
...
    CreateRuleSetRequest:
      type: object
      additionalProperties: false
      required: [key, name]
      properties:
        key: {type: string, pattern: '^[a-z][a-z0-9-]{2,62}$'}
        name: {type: string, minLength: 1, maxLength: 120}
    CreateRuleSetVersionRequest:
      type: object
      additionalProperties: false
      required: [yaml]
      properties:
        yaml: {type: string, minLength: 1, maxLength: 1048576}
    RuleSet:
      type: object
      additionalProperties: false
      required: [id, projectId, repositoryId, key, name, createdBy, createdAt]
      properties:
        id: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        repositoryId: {type: string, format: uuid}
        key: {type: string}
        name: {type: string}
        createdBy: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    RuleSetVersion:
      type: object
      additionalProperties: false
      required: [id, ruleSetId, version, yaml, sha256, scannerVersion, schemaVersion, createdBy, createdAt]
      properties:
        id: {type: string, format: uuid}
        ruleSetId: {type: string, format: uuid}
        version: {type: integer, minimum: 1}
        yaml: {type: string}
        sha256: {$ref: '#/components/schemas/Sha256'}
        scannerVersion: {type: string}
        schemaVersion: {type: string, const: 0.1.0}
        createdBy: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    SubmitScanJobRequest:
      type: object
      additionalProperties: false
      required: [repositoryId, ruleSetVersionId]
      properties:
        repositoryId: {type: string, format: uuid}
        ruleSetVersionId: {type: string, format: uuid}
    ScanJob:
      type: object
      additionalProperties: false
      required: [id, projectId, repositoryId, ruleSetVersionId, status, createdBy, createdAt, attempt, version]
      properties:
        id: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        repositoryId: {type: string, format: uuid}
        ruleSetVersionId: {type: string, format: uuid}
        status: {type: string, enum: [QUEUED, RUNNING, CANCEL_REQUESTED, SUCCEEDED, FAILED, CANCELLED]}
        outcome: {type: [string, 'null'], enum: [PASS, VIOLATION, null]}
        createdBy: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        completedAt: {type: [string, 'null'], format: date-time}
        attempt: {type: integer, minimum: 0}
        version: {type: integer, format: int64, minimum: 0}
        failureCode: {type: [string, 'null']}
        failureMessage: {type: [string, 'null']}
        reportSha256:
          oneOf:
            - {$ref: '#/components/schemas/Sha256'}
            - {type: 'null'}
    SourceLocation:
      type: object
      additionalProperties: false
      required: [path, startLine, startColumn, endLine, endColumn]
      properties:
        path: {type: string}
        startLine: {type: integer, minimum: 1}
        startColumn: {type: integer, minimum: 1}
        endLine: {type: integer, minimum: 1}
        endColumn: {type: integer, minimum: 1}
    Finding:
      type: object
      additionalProperties: false
      required: [id, projectId, jobId, scannerFindingId, fingerprint, ruleId, ruleVersion, severity, subjectId, message, disposition, version, evidenceIds]
      properties:
        id: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        jobId: {type: string, format: uuid}
        scannerFindingId: {type: string}
        fingerprint: {$ref: '#/components/schemas/Sha256'}
        ruleId: {type: string}
        ruleVersion: {type: string}
        severity: {type: string, enum: [low, medium, high, critical]}
        subjectId: {type: string}
        message: {type: string}
        location:
          oneOf:
            - {$ref: '#/components/schemas/SourceLocation'}
            - {type: 'null'}
        disposition: {$ref: '#/components/schemas/FindingDisposition'}
        version: {type: integer, format: int64, minimum: 0}
        evidenceIds:
          type: array
          items: {type: string, format: uuid}
    Evidence:
      type: object
      additionalProperties: false
      required: [id, projectId, jobId, scannerEvidenceId, kind, summary, location]
      properties:
        id: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        jobId: {type: string, format: uuid}
        scannerEvidenceId: {type: string}
        kind: {type: string}
        summary: {type: string}
        location: {$ref: '#/components/schemas/SourceLocation'}
    FindingDisposition:
      type: string
      enum: [OPEN, FALSE_POSITIVE, RISK_ACCEPTED]
    SetDispositionRequest:
      type: object
      additionalProperties: false
      required: [disposition, reason, version]
      properties:
        disposition: {$ref: '#/components/schemas/FindingDisposition'}
        reason: {type: string, minLength: 1, maxLength: 1000}
        version: {type: integer, format: int64, minimum: 0}
    Disposition:
      type: object
      additionalProperties: false
      required: [id, findingId, previousDisposition, newDisposition, reason, actorId, occurredAt, version]
      properties:
        id: {type: string, format: uuid}
        findingId: {type: string, format: uuid}
        previousDisposition: {$ref: '#/components/schemas/FindingDisposition'}
        newDisposition: {$ref: '#/components/schemas/FindingDisposition'}
        reason: {type: string}
        actorId: {type: string, format: uuid}
        occurredAt: {type: string, format: date-time}
        version: {type: integer, format: int64, minimum: 1}
    Sha256:
      type: string
      pattern: '^[0-9a-f]{64}$'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, traceId, details]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
          pattern: '^[0-9a-f]{32}$'
        details:
          type: object
          additionalProperties: true